+++ src/plugins/index.ts
@@ -1,2 +1,4 @@
 import tablePlugin from "./tablePlugin";
-export default [tablePlugin];
+import imagePlugin from "./imagePlugin";
+
+export default [tablePlugin, imagePlugin];
