+++ src/schema.ts
@@ -204,9 +204,5 @@ export default {
       },
     },
   },
-  inlines: {
-    image: {
-      isVoid: true,
-    },
-  },
+  inlines: {},
 };
